hfield fix implicit conversion

+++ include/hpp/fcl/internal/traversal_node_hfield_shape.h
@@ -223,13 +223,13 @@ public:
   /// @brief Obtain the left child of BV node in the first BVH
   int getFirstLeftChild(unsigned int b) const
   {
-    return model1->getBV(b).leftChild();
+    return static_cast<int>(model1->getBV(b).leftChild());
   }
 
   /// @brief Obtain the right child of BV node in the first BVH
   int getFirstRightChild(unsigned int b) const
   {
-    return model1->getBV(b).rightChild();
+    return static_cast<int>(model1->getBV(b).rightChild());
   }
 
   /// @brief BV culling test in one BVTT node
